Debugging, Creating Session Variables
Did some debugging, mostly by adding session variables and hidden inputs. So far, the table with show when submitted at the top, but the queries are kind of weird, especially not working with strings and with the output not working correctly. It is probably just a weird behavior when updating the table (hoping, praying even). Also added some skeleton code to kind of show the structure.